Deprecated checkSensors and sendData, renamed checkSensors to snakecase and combined sendData into server_interactions with the main loop sent to main.py
